client: merge generic response and error handling
Also fix a small bug in error handling (when unmarshaling the JSON, Go
may just return an empty list instead of a nil pointer, so we should
actually check the size of the error list). In order to avoid having to
do a errResponse.Errors != nil, I made Errors a non-pointer field.