encodingcom: implement GetStatus
This method will collect the current status of media being transcoded.
The whole code might seem as a hack, but I wanted to workaround the way
that Encoding.com reports the status of each destination, so I decided
to create this DestinationStatus struct, and in order to do that I
needed to replicate part of some structs, but we have tests for those
duplications! haha

Maybe we could take a simpler approach and just use two slices to
represent destinations and their status, but I think that it's a fair
trade-off to have this kind of ugly code hidden in unexported types and
do some make-up on before returning them to the caller code.

Closes #1.