encodingcom: fix GetStatus method
Encoding.com API doesn't behave as expected on when an object is a list.
When there's a single item in the list, it returns a JSON object, but
when there are multiple items, it returns a JSON list.

The solution is ugly, but it works: we first unmarshal the response in
to an interface{}, and then marshal the value back to a slice of bytes,
it also tries to convert this interface{} to a slice of empty interfaces
(i.e. []interface{}), and takes the following steps, according to the
result of the conversion:

 - if it can be converted to a slice of empty interfaces, then it's a
   JSON list, then we unmarshal the remarshaled slice of bytes into the
   slice of statusJSON
 - otherwise it's a JSON object, then we unmarshal the remarshaled sice
   of bytes into a single statusJSON instance and append it to the slice
   of statusJSON

The same applies to the slice of formats in each job, and for the slices
of destinations. Sounds like this API is just nuts.